move tests to better place
